+++ nvtabular/dispatch.py
@@ -1,0 +1,31 @@
+#
+# Copyright (c) 2021, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from typing import Union
+
+import cudf
+import pandas as pd
+
+DataFrameType = Union[pd.DataFrame, cudf.DataFrame]
+
+
+def _concat_columns(args: list):
+    """Dispatch function to concatenate DataFrames with axis=1"""
+    if len(args) == 1:
+        return args[0]
+    else:
+        _lib = cudf if isinstance(args[0], cudf.DataFrame) else pd
+        return _lib.concat(args, axis=1)
+    return None
